+++ core/include/scipp/core/dataset.h
@@ -963,33 +963,6 @@ SCIPP_CORE_EXPORT Dataset histogram(const Dataset &dataset, const Dim &dim);
 
 SCIPP_CORE_EXPORT Dataset merge(const DatasetConstProxy &a,
                                 const DatasetConstProxy &b);
-namespace detail {
-template <class DS, class Func>
-Dataset apply_through_dimension(const DS &ds, const Dim dimension, Func func) {
-  if (containsSparse(ds))
-    throw std::logic_error("Can't sum Dataset with sparse data");
-  if (!ds.dimensions().count(dimension))
-    throw std::logic_error("Can't sum Dataset on non existing dimension.");
-  Dataset res;
-  for (auto && [ name, attr ] : ds.attrs())
-    res.setAttr(std::string(name), attr);
-  for (auto && [ dim, coord ] : ds.coords()) {
-    if (dimension != dim)
-      res.setCoord(dim, coord);
-  }
-  for (auto && [ name, label ] : ds.labels()) {
-    if (label.dims().inner() != dimension)
-      res.setLabels(std::string(name), label);
-  }
-  for (auto && [ name, item ] : ds) {
-    if (!item.data().dims().contains(dimension))
-      res.setData(name, item.data());
-    else
-      res.setData(std::string(name), func(item.data(), dimension));
-  }
-  return res;
-}
-} // namespace detail
 
 [[nodiscard]] bool containsSparse(const DatasetConstProxy &ds) noexcept;
 SCIPP_CORE_EXPORT Dataset sum(const DatasetConstProxy &ds, const Dim dimension);
